Allow fractional counts.
git-svn-id: https://ws10smt.googlecode.com/svn/trunk@337 ec762483-ff6d-05da-a07a-a48fb63a330f

+++ gi/posterior-regularisation/prjava/src/phrase/Agree.java
@@ -112,7 +112,7 @@ public class Agree {
 				bdist += edge.getCount() * Math.log(z);
 				arr.F.l1normalize(p);
 				
-				int count = edge.getCount();
+				double count = edge.getCount();
 				//increment expected count
 				TIntArrayList phraseToks = edge.getPhrase();
 				TIntArrayList contextToks = edge.getContext();
